TP2/requirements.txt
aiohttp
beautifulsoup4
lxml
playwright
pillow
requests

